• Refactors and cleans up some of the code for homepages sections
