+++ anime-archive/src/components/animeCard/AnimeCard.js
@@ -1,7 +1,7 @@
 import "./AnimeCard.css";
 
 const AnimeCard = (props) => {
-  console.log(props);
+  const data = props.data;
   const month = {
     1: "Jan",
     2: "Feb",
@@ -17,7 +17,7 @@ const AnimeCard = (props) => {
     12: "Dec",
   };
 
-  const newMonth = month[props.data.startDate.month];
+  const newMonth = month[data.startDate.month];
 
   // Above object is being used to turn integer month returned from api into a string month
   // Then saving it as variable to use for start date in card below
@@ -25,16 +25,16 @@ const AnimeCard = (props) => {
   return (
     <div className="animeCard">
       <div className="cardImage">
-        <img src={props.data.coverImage.large} alt="anime cover" />
+        <img src={data.coverImage.large} alt="anime cover" />
       </div>
       <div className="cardDetails">
-        <h3 className="animeCardTitle">{props.data.title.english}</h3>
+        <h3 className="animeCardTitle">{data.title.english}</h3>
         <div>
           <h4>Start Date</h4>
-          <p>{`${newMonth} ${props.data.startDate.day}, ${props.data.startDate.year}`}</p>
+          <p>{`${newMonth} ${data.startDate.day}, ${data.startDate.year}`}</p>
           <br />
           <h4>Status</h4>
-          <p>{props.data.status}</p>
+          <p>{data.status}</p>
         </div>
       </div>
     </div>
