+++ anime-archive/src/components/section/Section.js
@@ -2,8 +2,8 @@ import AnimeCard from "../../components/animeCard/AnimeCard.js";
 import "./Section.css";
 
 const Section = (props) => {
-  const data = props.apiData.data.data;
-  console.log(data, props.data.extension);
+  const sections = props.apiData.data.data;
+
   // console.log(data[props.data.extension], "props in section");
   return (
     <section>
@@ -13,7 +13,7 @@ const Section = (props) => {
         <p>view more</p>
       </div>
       {/* With query we can pull 4 cards and cycle through dynamically */}
-      {data[props.data.extension].media.map((item) => (
+      {sections[props.data.extension].media.map((item) => (
         <AnimeCard key={item.id} data={item} />
       ))}
     </section>
