+++ anime-archive/src/graphql/index.js
@@ -1,0 +1,3 @@
+import { fetchHomepage } from "./fetchHomepage.js";
+
+export { fetchHomepage };
